Update COSE spec (#179)
* Update signature-envelope-cose.md (#153)

The `crit` header should be optional, not required. The spec already defines `cty` and `signingtime` as required, so a compliant verifier should reject any signature that does not include them, irrespective of whether or not the envelope also declares them as critical.

An appropriate use of `crit` would be if a signer intends to include a new header not defined in the spec, to instruct verifiers that they must understand and process that header in order for the signature to be successfully validated. Another scenario is if a signer intends to mandate processing of an otherwise optional header (e.g., `expiry`).

In other words, `crit` can be used by a signer to add requirements to the verification flow, but it is unnecessary and redundant if the verifier already has the same requirements.

Signed-off-by: Brian Krell <briankr@microsoft.com>

* `cose-envelope`: merge changes from `main` (#177)

* Align COSE spec with JWS

Signed-off-by: letmaik <maik.riechert@microsoft.com>
Co-authored-by: Steve Lasker <stevenlasker@hotmail.com>
Co-authored-by: Quim Muntal <quimmuntal@gmail.com>
Co-authored-by: Quim Muntal <qmuntaldiaz@microsoft.com>
Co-authored-by: Milind Gokarn <milind81@gmail.com>
Co-authored-by: Milind Gokarn <gokarnm@amazon.com>
Co-authored-by: Shiwei Zhang <shizh@microsoft.com>
Co-authored-by: Pritesh Bandi <priteshbandi@gmail.com>
Co-authored-by: Sajay Antony <1821104+sajayantony@users.noreply.github.com>

* update CBOR spec

Co-authored-by: Steve Lasker <stevenlasker@hotmail.com>
Co-authored-by: qmuntal <qmuntaldiaz@microsoft.com>
Co-authored-by: Milind Gokarn <gokarnm@amazon.com>
Co-authored-by: Pritesh Bandi <pritesb@amazon.com>
Co-authored-by: Sajay Antony <sajaya@microsoft.com>
Signed-off-by: Shiwei Zhang <shizh@microsoft.com>

Co-authored-by: Brian Krell <57008892+briankr-ms@users.noreply.github.com>
Co-authored-by: Maik Riechert <maik.riechert@arcor.de>
Co-authored-by: Steve Lasker <stevenlasker@hotmail.com>
Co-authored-by: Quim Muntal <quimmuntal@gmail.com>
Co-authored-by: Quim Muntal <qmuntaldiaz@microsoft.com>
Co-authored-by: Milind Gokarn <milind81@gmail.com>
Co-authored-by: Milind Gokarn <gokarnm@amazon.com>
Co-authored-by: Pritesh Bandi <priteshbandi@gmail.com>
Co-authored-by: Sajay Antony <1821104+sajayantony@users.noreply.github.com>
Co-authored-by: Pritesh Bandi <pritesb@amazon.com>
Co-authored-by: Sajay Antony <sajaya@microsoft.com>